FAI-1946 saved searches on home page

+++ src/SFA.DAS.FAA.Domain/Models/SearchParameters.cs
@@ -3,7 +3,7 @@
 public record SearchParameters(
     string? SearchTerm,
     List<int>? SelectedRouteIds,
-    decimal? Distance,
+    int? Distance,
     bool DisabilityConfident,
     List<int>? SelectedLevelIds,
     string? Location
